Add beacon support to WiFi
This change adds a basic beacon to our wifi module.  The beacon is
a json object with a beacon tag.  In it is the RCP type, serial,
ip list (it may have more than one IP bound to it), and port.  The
beacon broadcasts to address 255.255.255.255 (local network broadcast
address per RFC 919) and port 7223 on guests.