Add support for handling unhandled URCs in a generic fashion
So the AT state machine supports handling URCs via a registration
method.  However that design assumes that the AT URCs are in fact
sane to begin with.  Some less sane / rougher AT URCs just speak
their mind bluntly on the serial connection.  When this happens,
it can be hard to handle the message as a registered URC, mainly
because all registered URCs begin with a command prefix.  Thus this
patch introduces a way to handle these unclean URCs by introducing
a optional catchall that a programmer can use to handle these URCs.
This allows more flexibility without introducing an unreasonable
URC handlers.  Think of this as a last resort.

For a good example of why this is useful, see the ESP8266 module and
how it notifies the host device about when a connection is opened or
closed or when it connects/drops a WiFi network.