Properly check for mailbox on no timeout CAN tx
If a user called CAN Tx and specified a timeout value of 0, then we
would always return true.  This was bad beacuse its possible that
the message never made it into a mailbox, and thus will never be sent.
This trivial patch checks the return value of the CAN tx value so that
we return true if a mailbox was obtained, and false if none were available.
This will enable stronger confidence in our CAN Tx abilities.

Fixes Issue #549