Re-tool the state machine in our esp8266_drv
So as I continued to add more features to this driver it quickly
became apparent that my original design was insufficient.  The main
issue was that decision making code was placed in callbacks.  This
made it very hard to get the state machines (there are several) to
play well with one another.  So that has all changed around.

Now what we do is introduce a mechanism that allows us to "flag"
various state machines that they need to check thier stte.  Combined
with a priority mechanism this design allows us to put all the
decision making logic into one method for each state machine (or a set
of methods all with a single entry point like with check_client).  The
callbacks are responsible now for simply updating the state and for
setting flags that tell each state machine that it needs to check
itself.  This design greatly simplifies the logic, putting all of it
in one central location.  The result is a much faster and more
efficient state machine that is easier to follow.

I also cleaned up some annoying things like the private vars that
were prefixed with '_' characters.  That was unnecessary and I have
learned my lesson.