Fix up our serial to behave a bit more consistenly and deal with fallout
Our serial_buffer code used to do some not so good things like strip
off control characters.  While at first this was good I have learned
you want to do this higher up because sometimes you need to view those
characters.  So this change adjusts serial_buffer to not strip out the
control characters.  It will continue to ignore serial commands with
no useful data (ie a line ending in "\r\n") but that is it.  Commands
higher up are responsible for dealing with the trailing characters.  Its
trivial to strip them using the commands in str_util.

This also fixes bugs in our serial_read_line code where we might not
properly terminate strings etc.  Now we don't handle that at that layer
and instead deal with it higher up when we are sure we want to do it.