Backport the MK2 ORE fix to 2.9.1
As explained in PR #561 there was a bug with ORE resetting in MK2
where we were only resetting USART1 instead of resetting the handle
that was causing the ORE to occur.  This change fixes that by using
the provided handle to do the steps needed to clear the ORE flag.