Improve the rx_buff, improve wifi task rx capabilities
This change improves the rx_buff by making the changes necessary
for it to be fast, lean, and consistent.  This buffer is used
for receiving command messages, and is designed to replace a lot
of the duplicated code we have spread in the firmware.  Its designed
to recieve partial messages or complete ones, allowing us to quickly
cycle through Serial streams to ensure we block as little as possible.

The code also gracefully handles overflows along with provides
detailed state to allow us to make the best decision based on what
data we have (or have not) received.

This change also cleans up the Wifi task code a bit to be a bit
more friendsly towards commands and not wait for new commands that
may come in unless one is already partially in progess.