Add softAP support to our ESP8266 driver
Adds support for managing and controlling the soft AP on
the ESP8266 device.  This patch adds the ability to set and control
the soft AP bits in the wifi driver only.  A subsequent patch will
add the APIs necessary and will integrate with our LoggerConfig
structure for persistent configuration.

This patch does not add the ability to disable the softAP.  That
issue will be fixed by issue #597