Replace all `#if defined(...)` macros with ones that are always defined
Turns out that using the existence of a macro as a test can lead to some
tricky situations when you need to use the value of the macro as a boolean.
So instead of having some of the most hidious code ever seen by man, I opted
to fix the root of the problem and force the definition of all of our macros.
This way our builds will blow up if a macro is not defined properly someplace.
This directly contradicts my earlier commit making it the way it was... live
and learn I suppose since I am not perfect.