Make our SoftAP active by default with a unique SSID
Because there is a chance users will not have a USB or computer
around to connect to the RaceCapture for the first time, we need
a way for them to get into it.  This change enables that opportunity
by establishing a relatively unique SSID for RaceCapture to use based
on its cpu serial number.  We set this with no password for the default
network so that users can get into their device.

Yes this is not secure. Its not meant to be unless the user configures
it to be secure, so don't get me started on that track.