Retool how we handle callbacks on esp8266 driver and add functionality
This change re-tools how we handle registering callbacks against the
esp8266 driver.  We now register all callbacks using a single call and
pass a struct in with the pointers.  This is both flexible and less
invasive on the API.  This also centralizes where this action happens
instead of happening in two different spots like we previously did.

This change also reduces the number of callbacks by making the
callbacks more general.  Instead of doing specific actions, the
callbacks are invoked when general state change occurs, and relevant
information is passed in as part of the parameters.  This reduces the
number of callbacks and reduces a little bit of code.

Finally this change also handles some previously unhandled callbacks
that are only seen during driver init.  Now we act upon them to ensure
our state is always up to date.

Issues: #599, #585