Implement one IP retrieval call to rule them all
Turns out the same call will return both the station IP and the
client IP.  So adapt the call to provide both values and also
ensure that we feed that info into the correct spots and cause
the appropriate state machines to update.

This change also moves IP information into its own struct like
it should have been from the get go.  This is cleaner and works
better than stuffing it into the client info like we did before.

Issues: #604 #598