Make all non-tagged builds DEBUG builds.
We tag all RC and Release build for RaceCapture.  To ensure that we
find bugs more quickly we should always build with the watchdog
disabled so that bugs that cause crashes are more noticable and thus
get squashed sooner.  This change does that by enabling the DEBUG flag
on all non-tagged builds.  Tested a tagged build to ensure that it had
no DEBUG flag set.