Enable CHAP support for APNs that use PPP auth
For networks that make use of secure auth (PPP based) we should
support them.  To best handle this we enable CHAP support whenever
a user supplies a password in the APN field.  Supplying a password
implies that there is actually some form of secure authentication
happening, thus we should enable an authentication mechanism on our
APN settings.

For most providers here in the states, there will be no change since
none of them give a password for the APN.  But for some of our friends
overseas, this will be useful.

Note this only applies to our 3G chips.  2G procedures will not change.

Issue #619