Add location for common CFLAGS.  Use common CFLAGS
This patch defines the asl_cflags.mk file.  This is where we keep
common flags that apply to all projects.  The top level Makefile
includes these flags and exports them so that all projects see them.

The benefit is we now no longer need to worry about defining the same
common flags in 3 different areas.  Instead we define them in a single
location.