Implement WiFi Mode control
The ESP8266 module has various modes that are used to enable/disable
various functionality.  We need to be able to set these modes based
on the user configs so the device acts correctly.

This patch adds support for controlling the mode and responding to user
changes as they come in.

Issue: #597