Improve tracking of how much data was actually sent
There are cases currently in our send data code where we might think
we sent no data but actually did.  The inverse was also true.  This
patch fixes that by tracking how much data was actually removed from
the serial buffer and subsequently subtracts it from the buffer len
counter.  This ensures that these values are kept in sync.