Move to address based DMA
In the original DMA impelentation I was using a sentinal character
to inform me of when I had run out of data. This was not very clean
and was only done because I thought I had no way to figure out where
the head of the DMA pointer is.  However I found there is a way to
calculate it.  So this change removes the use of the sentinal character
and instead uses addresses like a sane program.