Get GPS half-working on the new DMA system
Move to using the new DMA code from the previous DMA implementation
for GPS Rx buffer.  This new implementation allows for a smaller buffer
to be used. I say half-working because I have disabled the reading of
the DMA data during the timer expiration.  For some reason this is
causing the GPS code to fail disabling NMEA.