Poison our Data struct to find errors; Fix erros found
I was getting different behavior in code due to a somewhat related
but completely different piece of code changing. So to figure out
if something was not getting written that was being sent, I wrote
poisoning code to poison the structure with a known pattern so it
would be easier to spot the bug.

Once the bug was spotted, I was able to quickly isoloate the
root cause (a copy/pasta error) and fix it.  I also verified the
other structures were correct.