Fix GPS responding too quickly bug, demote trace log msgs to proper level
This patch fixes an issue I encountered with the DMA enhancement to RCT.
Turns out that you can in fact respond to the GPS unit too quickly.  When
this happens the unit will not provide an ack.  This screws up our code.
The quick and dirty fix to this is to add a back-off time before we send
a message.  The lowest time value that I found to be stable was 10ms after
an ack was received.  So that is what this patch primarily does

This patch also cleans up some of the debugging so that it is logged at
the trace level instead of the info.  This ensures we keep our log clean
of spurrious messages.