Backport the improved GPS code from RCT to MK2.
Backports the fixes made for the GPS code in RCT to the MK2 branch.
This will be further improved when we extract common code in a later
fix. That is tracked by issue #643

Also ups the backoff timeout to 20ms.  Per issue #630
Also removes the poison portion of the code since this is not necessary.