Improve release package naming, move RCT bin files to top level
Improves the release naming of our packages by using the output of
`git describe --dirty` to name the package.  This always gives us
a unique version that we can give to customers _AT ANY TIME_.  This
way we know EXACTLY what version of the product they are using because
the version now has all of this information.  And when it comes to
official and RC releases, the action of tagging will make our version
appear exactly like it used to.

This patch also enables us to start building .zip packages for the
RCT project in anticipation of delivery.  It also improves our
clean target and improves and sorts the contents of .gitignore so
we can quickly find what is in there (and removes duplicate entries).

Issue: #495