Up BT baud to 230400, remove unnecessary baud from config
This change ups the default BT baud rate to 230400, effectively
doubling the rate at which we can send data over BT.  Combined with
the DMA improvements the result is a snappier BT experience.

This change also removes the baudRate value from the bluetooth
config since that was not being used in any significant way.

Issue #530