Support the "ALREADY CONNECT" AT message
This message gets returned when we try to open up a connection to
the same destination that already exists on this channel. We need to
treat it as a success message since no action was taken on behalf of
the esp8266 device yet since the connection is already present.

It also means we need to chance how we respond to connect messages
since the "ALREADY CONNECT" message does not supply any comm channel
information.  Nice of em eh?