Don't wait for data queues in the Rx handler
So originally I was trying to give the system time to catch up and
process data that was waiting in the queue.  However it would appear
that this is not reasonable since the driver keeps getting wedged in
this Rx loop.  So now we just drop data if it can't fit in the queue.