Don't always drop data if incomming buffer is full
Give the system a chance to context switch back and drain the buffer
if it can.  If it can't then we are boned anyways so we have to
drop it.  This seems like the best comprimise here to ensure that
the RX of data has a chance while keeping our buffers small for
RCT.