Finish up implementation of sample stream & reorganize
The main portion of this commit compeletes the wifi streaming
API setup by completing the implementation end to end.

Because there are some testing issues with what files are
compiled during test I needed to migrate and rename some of the
streaming code.  Hence why lots of changes in loggerSampleData.

Issue #515