Add WiFi TxRx light support to RCT
This change adds a nice and easy to see LED light.  Its activated
when we send or receive data over any socket.  It does not turn on
for AT commands.  I chose this route so customers would be able to
know if the wifi unit was communicating with the outside world or
not.  Also its consistent with how our external modules blink.

NOTE: The LED only affects RCT and not MK2.  Why?  Because there are
no specific WiFi lights on MK2 other than the external module light.
And that one seems to work the same way (blinks only when data is
sent out).

Issue #621