Make WiFi LED use a one-shot timer; trigger on AT commands.
Based on good feedback from Jeff I re-implemented this logic so that
we only fire the timer as needed instead of firing it _every_ period
time.  This will save on some resources and is generally a better
design.  The tradeoff was one extra variable in our state structure.
Totally worth it.

Also we now flash the WiFi LED based on AT commands being sent.
Originally I stated that the LED on the ESP8266 module from SparkFun
only blinked when data was sent. This wass incorrect. Instead it blinks
when any Serial data is sent (including AT commands). This
patch corrects that misstep by blinking the LED during the completion of
any command and when new data comes in.