Move our distance calculation to one that is speed/time based
As we have seen in the field measuring distance by taking the
pythag distance between our current point and the last is not good
in many situations due to GPS inaccuracies.  Speed over time may
be a better approach since we have noticed that speed calculations
tend to be very accurate relative to the sometimes erratic location
calculations.  Hence this change.

Issue #178