Loosen our restriction on strictness of GPS task
This change loosens how strict our GPS subsys is when it comes to
reading GPS messages.  Instead on failing on 1 consecutive message, we
instead now fail when the number of failures is > then # of successes
and we cross the boundary of 3 total failures.  This allows us to miss
a couple of GPS messages without triggering a re-enum, which is probably
not necessary.