Compute sha256sum of main.* files as part of full build, clean .ihex files
Because we want to easily be able to check if our outputs are the same as
our inputs this change calculates the hashes of our main.* files under
platform to ensure that they are what we expect them to be.

This change also causes us to delte .ihex files when we run make clean.