Add support for valid track info to status message
We need to be able to know if the track that has been configured is
actually valid.  In example, if a user has a user defined track set
but its not valid (no valid start/finish), then they will not get any
lap data. Same situation could happen if auto-detect of a track fails.

This change adds a "valid" field to the track status portion of the
status endpoint.  By adding this value we are able to know with certainty
whether or not the user has a workable track configured.