Fix and Improve ADC setup on RCT
ADC wasn't firing more than once after init. We had properly setup
the init structure for the device but had missed setting the
ADCDmaConfig to circular (defaults to one shot).  This is what was
causing the failure.

In the process of doing this I read most of the reference manual on
this device and implemented some other bits we will benefit from
(like calibration).  I also cleaned out a lot of unnecessary fluff
since most of it was default settings.

Issue #520