Work past invalid ISR offset value in CPU init code
For some strange reason the linker script is returning an offset
value of 0x20020000 instead of the expected 0x00020000. If it weren't
for some good code on STM's part, this would blow up.  I don't know
why the linker is doing this but to work around the issue I have
masked the result with 0x000FFFFF.  This should get us sane values.