Fix out of bounds writing in init_cpu_id method
The code was wildly exceeding its boundary limits, corrupting other
variables in the .bss section around it.  We didn't notice this because
this method was the first thing to run and would not alter values after
that point.  However this overwriting caused corruption in the FatFS
subsystem which lead to a crash and issue #714.  This patch corrects
that behavior and implements simpler logic.

Issue #714.