Don't error on missing `onTick` method in Lua
This change alters our Lua subsystem to no longer error on a
missing `onTick` method.  This will allow users to create scripts
that run one time during first load of the script and will also
set us up to be able to fire callbacks in future versions should
we choose to do so.

Issue #707