Add flags support to "getCapabilities" API
Flags tell the caller what features will be supported on the unit.
This allows us to selectively enable/disable windows in the app so
that the user only sees what they need to see.  This also allows us
to have the app be flexible as we add new features going forward while
retaining backwards compatibility as much as possible.

Issue #709