Add auto logger subsystem
Auto Logger allows you to start/stop logging automatically based on
pre-defined speed and time values that are user adjustable.  For
starting simply set the minimum speed to maintain and for how long it
should be maintained. For stopping set the maximum speed that should
can be obtained and the time we must be underneath that speed.  Either
can be disabled by setting a value of 0 for the time value.  Time is
measured in seconds and speed in MPH.

Issue #169