Add logic to prevent user contention; change stop default
Adds logic so that we will try to not fight the user actions when
they push the button to either start or stop logging manually.
Without it the two can be in direct competition, and that is no
good.

Also adjusts the default values so that we stop logging sooner.  This
helps us not be writing to SD when the unit looses power, thus helping
to prevent SD card failure.