Add support for flexible boolean in Lua
In some of our Lua API calls we are using 0 and 1 values where we
should really be using false and true values.  These include

* setGpio
* setLed

This patch adds some new methods to allow users to either specify a
boolean parameter or use the legacy behavior and specify a digit to be
interpreted as a boolean.  0 is false and everything else is true.

Issue #628