Add support for non read modify write operations to flash
In cases where it is unreasonable to load the entire DB of a particular
object into RAM we need to be able to write the contents to flash
directly so as to avoid the large/impossible allocation. This change
adds that support so that we may write to flash without the need of
allocating such a ridculously large buffer.

Issue #536