Move to CGREG command from CREG command
After in-depth conversation with our friends at U-Blox we learned
that CREG is a registration command that does more than just GPRS.
It also covers registration of voice services. While this is ok in
many cases it likely will not work with some data-only sims because
they have no voice capabilities and the MNO may configure the SIM to
disallow any registration of voice services (which is logical).

So instead of checking if voice services are registered, this command
will only check if GPRS services are registered.  From there if we
are registered then we can activate a PDP context through the APN.
Hopefully this will add the ability to use more data-only sims as
they become more prevelant.

Issue #742