Ensure lapstats channel sample rates are consistent
As described by issue #503 some users are sometimes getting inconsistent
sample rate settings in their configs.  This is because older configs
sometimes have these values being separately configurable and thus
can introduce the bug.  While this is technically a bug in the app it
is also a firmware bug since we have yet to change the API to enforce
consistent behavior.  That issue is tracked by issue #748.

Issue #503