Fix CAN masking capabilities on MK2 and RCT
Turns out that our driver for setting the CAN ID message masks was
completely and totally incorrect on the stm32f4 and f3 chips.  This
change addresses the issue and fixes them so they work properly.

This change also goes farther to re-initialize all filter values to
0 upon CAN channel init and to set one filter to accept all messages.
It also exposes the ability to selectively enable/disable a filter.

Issue #737