Expose 3 GPS channel units to API to allow for metric support
This change exposes the following GPS channel unit label values to the
logger API with the following supported values (default first):

* altitude - ft,  m
* distance - mi,  km
* speed    - mph, kph

Unlike other unit labels though this is the first to enforce strict
values. In this case there are only two options with the first label
being the default should the other option or an invalid option be
specified.  In case the labels aren't clear they are as follows:

* ft  - feet
* km  - kilometers
* kph - kilometers per hour
* m   - meters
* mi  - miles
* mph - miles per hour