Improve SD card writing by using new ring_buffer DMA methods
To help improve the stability of our SDCard code we now will use dma
methods that allow us to directly access the ring buffer.  This saves
us an intermediary call and allows us to pass in large blocks of data
to write into the f_write function.  This seems to make that method
happier and more stable.

I also increased the buffer from 512 to 1024 bytes, which also appears
to help.  I am seeing fewer write errors as a result.