Change serial get/put to read/write; update return signature
In order to sanely be able to close a serial device we must be able
to determine if the read/write operation failed.  We do this by
following the traditional file object approach and return -1 if
there was an error.  This means we had to change a few of the get
and put operation signatures.

In doing this I also took the opportunity to rename get/put to
read/write.  This allowed me to view all spots where these calls
are used and add appropriate checks as needed for this change.  It
also allows me to get a better handle on how large a project it will
be to migrate away from Serial to a more generic handle type once
we get there (hopefully in the next release or two).