Alter our wifi task to be responsible for managing Serial connections
It makes sense for the task that handles the Serial IOCTLS to be the
one that manages the connections.  This change effectively achieves
that by moving the responsibility of closing connections and dealing
with all associated fallout to the managing task (wifi.c in this case).

This allows us to shut off telemetry streams if a connection closes,
but does so in a way where we can ensure that we will not try to
act upon an invalid Serial device (because this task its the only
one that should destroy it).  Hence we find sanity in that and
do things a bit more properly.