Handle rate adjustment cases
Turns out my refactor had a bug where if you adjusted the rate it would
create a duplicate stream. So that is no good.  This change fixes that
by ensuring that if such a call happens, we first close the old stream
before creating a new one with the specified rate.

Issue #787