Validate LoggerMessage objects prior to writing
Its possible that our LoggerMessages may be expired by the time we
try to write them.  When this happens we may write stale data or even
no data, due to how our LoggerMessages currently work.  That is no
good.

This patch addresses that issue by first validating the LoggerMessage
structure to ensure that it is not stale, and then trying to write it.