Fix wifi init hang; add better init logging
We were encountering an overflow issue in our logic that would cause
the WiFi subsystem to hang and never recover during init if it was
enabled and no wifi device was attached. This patch fixes that and
adds lots more debug to help out with future issues.

I also added a lengh parameter instead of requiring a NULL terminated
list of bauds for the probe command.  I figured this was a saner
approach less prone to failure.

Issue #841