Work around the connection check flooding
Currently we use a polling mechanism on our wifi task to check for
connections.  We use this because there is no event driven method
that we can use (yet) to make this more efficient.  Without this
gate we will flood the event queue on the Wifi task, which in turn
will cause messages to be dropped.