Increase Rx Buffers to 768 bytes, make universal, Add support for adjustable serial buffers
We are crushing our Rx buffers when our app is programming new settings to RCT.
To counteract this I need to increase those buffers so that they may take the maximum
size of data we send them.  But then that leads to us running low on RAM. To counteract
this effect I have added support for the creation of variable serial buffer sizes.
I have also made the Rx buffer size a universal value that should be consistent across
the entire firmware.