Fix how we close our serial devices
Turns out that deleting a queue while a task is waiting on it is a bad
thing. I should have known better but in a moment of vulnerability I
was convinced to a silly thing like what I described above. But I have
learned from the experience and thus return stronger and with more
sanity.

I also improved how we flush queues by using the more optimized
xQueueReset method.

Issues #812 #772