Disable track auto-detect by default if trackDB size is 0
If we have no track DB, then auto-detect is useless.  So we should
disable it by default in that case. Otherwise we should keep the
traditional behavior of enabling it by default if we do have a
track DB available (whether or not it has tracks is another story).

Issue #814