Cleanup at_info_init, move device config into its own method
This change cleans up the at_info_init method by renaming it to be
consistent with our other public APIs.  It also moves all of the
device configuraiton to be done through the at_device_config method
instead of doing it both through the init and the device_config
method.