Handle no Serial EOF at end of ESP8266 IPD command
So the IPD command which is a callback for when we get data has no
terminating \r\n.  This is a problem because we rely on that usually to
tell our serial line parser that a line has been read.  We get lucky
since ESP8266 chip does follow some other AT standards and does send \r\n
before next command so we get some extra characters in there sometimes.
This change add checks to ensure we read as much as we expected to read
and clips off messages that are too long because of the bug.