Add WiFi "active" and "initialized" statuses
In order to be better able to see and evaluate user issuers in the
field we need to add some additional WiFi status information to our
status API.  This change adds both the active and initialized fields.
Active tells us whether or not the device is configured active in the
config and initialized tells us the driver's status.

Issue #829