Adjust WiFi client to better handle connect/disconnect actions
The behavior of the WiFi client on the ESP8266 was not entirely
clear and thus I ended up coding it incorrectly.  This change addresses
that so that we use the auto-connect features that are built into the
ESP8266 chip and better handle the callbacks that we get from it.

Issue #822