Move RCT .bss to CCM_RAM; Make heap larger
Moves the .bss segment to the CCM_RAM segment since it is now small enough
to fix thanks to the migration of the JSON token parser to the heap.  This
then makes the heap larger by splitting the remaining SRAM segment into a
DATA segment and a HEAP segment and placing the respective values in each.
The DATA segment is 4K while the HEAP is 36K.

Ran on RCT, works well and extra RAM shows.