Ignore Socket URCs during open/close; ease up on serial reaping
Turns out there was an unexpected side effect with the Rude AT support.
Whenever we were opening up a new socket, we were treating it as an
external connection because we were handling the "X,CONNECT" command
as a URC (Thank you very much ESP8266!).  So we have to handle that
specially by checking if we are doing an open or close op on the channel
in question.  If so, then we ignore the URC, else handle it as we normally
do.

This change also eases up on how aggressively we reap serial devices.
Now we will only reap them in the spot where they are created.  This
makes the most sense as it sets a single place for this to happen
which avoids race conditions on non-existent resources and also allows
us to have a single point where we can handle the ambigious states
(like if there was a sudden failure of a serial connection, yet we are
still using handles internally).  Doing this effectively fixes the crash
bug while leaking as few resources a possible.  We don't technically leak
any resources, just don't close as aggressively.

Combined with the memory increases, all multi-connect issues are gone from
what I can tell.

Issue #797