Fix panic flash counts
Because of how panic is coded, it appears that there are 4 flashes
when there are really 3.  This is not good as its not useful for
debugging.  To fix this we just iterate over the loop one less time.
Also had to adjust our ENUM values to account for the change and
introduce the "UNKNOWN" value to represent no flashes.

Issue #792