Add optional "apply" flag to setWifiCfg command
This change adds the "apply" flag to the setWifiCfg JSON command.  This flag, if
set to true, will apply the new configuration of the wifi settings until either
the next call with "apply" set to true, or until the next reboot (whichever happens
first).

In making this change, I had to alter the ESP8266 driver to keep a local copy of
the configuration.  This is so that user changes to the configuration in NVRAM
would not have any effect on the current ESP8266 wifi client or AP settings.

Issue #854