Run Lua interactive commands in Lua Task
This semi-complex change causes us to run interactive Lua commands
in the Lua Task as opposed to the task hosting the interactive Lua
session.  This is done because we don't know the stack depth of
the task hosting the interactive session, hence we can't gaurantee
it will not overflow when invoking Lua commands.

Issue: #844