Keep Lua runtime active, even if no code to run.
Because we want to continue to have support for an interactive
Lua session we should not kill the Lua task if no periodic function
is defined.  This along with the other changes around how interactive
sessions are managed will allow us to always have and use the Lua
environment (even if there is no script).