Properly encode/decode JSON strings in firmware
Turns out we were not properly encoding/decoding JSON strings.
Certain characters are considered illegal in a JSON stream and
thus must be encoded (escaped) so they become legal charaters.
They then must be unencoded on the receiving end so that their
meaning is restored.

This change adds in encode and decode methods, makes use of them
at the appropriate spots and removes broken code that previously
was doing this incorrectly/incompletely.

Note: We don't support unicode encoded charaters.  I just replace
those with a '?' if I see them.

Issue #808