Restrict what GPS points predictive timer will accept.
To help deal with situations where GPS quality drops, this change
adds a basic check to ensure that the GPS snapshots accepted by the
predictive timer are of high enough quality to reflect an accurate
time. If the sample is too low of quality, then the code will ignore
the sample.  This applies to both adding samples and retrieving
predictive times.

Issue: #108