test: add test to validate distance calculation
When reviewing my code for a presentation I thought that I might have
a bug in my distance calculation and thus wanted to validate that it
was there and needed fixing or that it was not and learn where I was
wrong. Turns out there is no bug when compared to an extermal tool
using a the Haversine method to calculate the distance. The only
concerning bit is that our calculation is off by about 0.5 meters at
75 meters distance.