fix: correct predictive timer interval calculation
We were dividing our slots value by 0.9 (90%) when we should have
been multiplying this value by 90%. We want to target usage of
90% of our slots and this is the correct math. Otherwise we will
constantly overflow and have poor results at the end of the lap
when we run out of buffer space.  This leads to poor measurements
and thus the jumps people were seeing.  *sigh*

Resolves: #108